requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain-core>=0.2.0

# PDF Processing
pypdf>=3.0.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
Pillow>=11.0.0

# HTTP & Async
httpx>=0.28.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=9.0.0

# CLI
click>=8.0.0
